Cleaning up

+++ include/trainer.h
@@ -2,4 +2,4 @@
 #include "nn.h"
 #include "plot.h"
 
-void tnrBatchTrain(NN* nn, NN* g, Mat* td, Plot* plot, size_t batchSize, size_t tiColSize, size_t toColSize, size_t rate);
+void tnrBatchTrain(NN* nn, NN* g, Mat* td, Plot* plot, size_t batchSize, size_t tiColSize, size_t toColSize, float rate);
